fix: rename models submodules
