+++ src/drift_benchmark/models/results.py
@@ -9,8 +9,8 @@ from typing import Any, List, Optional, Union
 import pandas as pd
 from pydantic import BaseModel, Field
 
-from .configuration_models import BenchmarkConfig
-from .metadata_models import BenchmarkSummary, DatasetMetadata
+from .configurations import BenchmarkConfig
+from .metadata import BenchmarkSummary, DatasetMetadata
 
 
 class DatasetResult(BaseModel):
